Parse gridfm datasets through the universal parse
The gridfm Parquet reader moves out of powerio-matrix into the facade,
which owns universal dispatch: a dataset directory now parses to
PioValue::BalancedNetworkScenarioSet through the retained directory
source, with the raw table prefix resolved over the entry listing, the
parquet tables decoded zero copy from acquired buffers, and each
scenario's read findings on the module. Matrix construction stays never
a parse result: powerio-matrix keeps the write side (which derives
y_bus_data and branch flows) behind its gridfm feature, and its round
trip tests read back through the facade. The facade takes arrow and
parquet as unconditional dependencies so automatic parsing does not
change with Cargo features; the compile cost lands on facade consumers,
who get the complete compiler the crate advertises.

The C ABI's pio_read_dir/pio_scenario_ids, the Python read_gridfm
functions, and the CLI read paths repoint to the facade reader with
unchanged behavior. prob_pyerr lost its last caller and goes.
