+++ powerio-matrix/src/io/gridfm.rs
@@ -67,16 +67,14 @@ use arrow::datatypes::{Field, Schema};
 use arrow::record_batch::RecordBatch;
 use num_complex::Complex64;
 
-use crate::diagnostics::codes;
 use parquet::arrow::ArrowWriter;
-use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
 use parquet::basic::Compression;
 use parquet::file::properties::WriterProperties;
 use serde::Serialize;
 
 use crate::indexed::IndexedNetwork;
 use crate::matrix::{BuildOptions, YbusFlags, branch_admittance, branch_flows, build_ybus};
-use crate::network::{Branch, Bus, BusId, BusType, Generator, Load, Shunt, SourceFormat};
+use crate::network::{BusId, BusType};
 use crate::{
     BalancedNetwork, ElementCounts, Error, GenCost, GenCostPatch, MissingGenCostPolicy, Result,
     ScenarioMismatch,
@@ -1077,750 +1075,17 @@ fn with_scenario_pair(
 // synthesized `1..n`, nodal load/shunt is folded to one synthetic element per
 // bus, and HVDC/storage/piecewise costs are gone (see [`GridfmRead::warnings`]).
 // `y_bus_data` is never read; branches carry raw `r/x/b`.
-
-/// One scenario read out of a gridfm dataset: the reconstructed [`BalancedNetwork`] plus
-/// the fidelity warnings the lossy read couldn't avoid (mirroring
-/// [`Conversion::warnings`](powerio_tx::Conversion)).
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub struct GridfmRead {
-    /// The reconstructed network (`source_format = SourceFormat::Gridfm`).
-    pub network: BalancedNetwork,
-    /// The scenario id these rows came from.
-    pub scenario: i64,
-    /// What the gridfm schema couldn't round-trip — synthesized bus ids, folded
-    /// per-bus load/shunt, dropped HVDC/storage, etc., as structured records.
-    pub diagnostics: Vec<powerio_core::Diagnostic>,
-    /// The same findings as `CODE: message` lines, rendered from
-    /// `diagnostics`.
-    pub warnings: Vec<String>,
-}
-
-/// Build one [`BalancedNetwork`] from in-memory gridfm tables, selecting `scenario`'s
-/// rows. The pure inverse of [`gridfm_record_batches_single`]: `base_mva` and `name`
-/// come from the caller (the disk path reads them from `gridfm_meta.json`).
-///
-/// # Errors
-/// [`powerio_tx::Error::FormatRead`] if a required column is missing or mistyped, a column
-/// carries nulls, a dense index is negative, or `scenario` isn't present; plus
-/// whatever [`BalancedNetwork::validate`] rejects (duplicate / dangling bus ids).
-pub fn read_gridfm_network(
-    tables: &GridfmTables,
-    scenario: i64,
-    base_mva: f64,
-    name: &str,
-) -> Result<GridfmRead> {
-    let bus = bus_columns(std::slice::from_ref(&tables.bus))?;
-    let gens = gen_columns(std::slice::from_ref(&tables.generator))?;
-    let branch = branch_columns(std::slice::from_ref(&tables.branch))?;
-    build_network_from_columns(
-        &bus,
-        &gens,
-        &branch,
-        scenario,
-        base_mva,
-        name,
-        crate::diagnostics::Diagnostics::new(),
-    )
-}
-
-/// Read one `scenario` from a gridfm dataset on disk and rebuild a [`BalancedNetwork`].
-/// The inverse of [`write_gridfm_dataset`].
-///
-/// `dir` is resolved leniently: the leaf `raw/` directory holding the parquet
-/// files, a `<case>/` directory with a `raw/` child, or a parent directory with
-/// exactly one `*/raw/` child all work. `base_mva` and the case name come from
-/// `gridfm_meta.json` (a missing manifest defaults `base_mva` to 100 and warns).
-///
-/// # Errors
-/// Propagates [`read_gridfm_network`] plus any filesystem / Parquet read error.
-pub fn read_gridfm_dataset(dir: impl AsRef<Path>, scenario: i64) -> Result<GridfmRead> {
-    let raw = resolve_raw_dir(dir.as_ref())?;
-    let (base_mva, name, warnings) = read_meta(&raw);
-    let bus = bus_columns(&read_parquet(&raw.join("bus_data.parquet"))?)?;
-    let gens = gen_columns(&read_parquet(&raw.join("gen_data.parquet"))?)?;
-    let branch = branch_columns(&read_parquet(&raw.join("branch_data.parquet"))?)?;
-    build_network_from_columns(&bus, &gens, &branch, scenario, base_mva, &name, warnings)
-}
-
-/// Read every scenario from a gridfm dataset, one [`BalancedNetwork`] per `scenario` id
-/// (sorted ascending) over the shared topology — the read side of the scenario
-/// batch (#57). Each scenario is rebuilt independently, so two scenarios may
-/// differ in branch status, bus types, and reference bus.
-///
-/// # Errors
-/// Propagates [`read_gridfm_dataset`]'s filesystem / Parquet / build errors.
-pub fn read_gridfm_scenarios(dir: impl AsRef<Path>) -> Result<Vec<GridfmRead>> {
-    let raw = resolve_raw_dir(dir.as_ref())?;
-    let (base_mva, name, warnings) = read_meta(&raw);
-    // Extract every column once and reuse across scenarios; rebuilding each
-    // scenario from the raw batches would re-concatenate each table n_scenarios
-    // times (O(n_scenarios × table_size)).
-    let bus = bus_columns(&read_parquet(&raw.join("bus_data.parquet"))?)?;
-    let gens = gen_columns(&read_parquet(&raw.join("gen_data.parquet"))?)?;
-    let branch = branch_columns(&read_parquet(&raw.join("branch_data.parquet"))?)?;
-
-    distinct_sorted(&bus.scenario)
-        .into_iter()
-        .map(|s| {
-            build_network_from_columns(&bus, &gens, &branch, s, base_mva, &name, warnings.clone())
-        })
-        .collect()
-}
-
-/// The distinct scenario ids in a gridfm dataset, ascending — the keys
-/// [`read_gridfm_scenarios`] rebuilds a [`BalancedNetwork`] for. Reads only `bus_data`'s
-/// scenario column, so it enumerates a dataset's scenarios without rebuilding
-/// every network; the C ABI's `pio_scenario_ids` is a thin wrapper over it.
-///
-/// # Errors
-/// Propagates the directory resolution and `bus_data.parquet` read errors.
-pub fn gridfm_scenario_ids(dir: impl AsRef<Path>) -> Result<Vec<i64>> {
-    let raw = resolve_raw_dir(dir.as_ref())?;
-    let bus = bus_columns(&read_parquet(&raw.join("bus_data.parquet"))?)?;
-    Ok(distinct_sorted(&bus.scenario))
-}
-
-/// The distinct values of `scenario`, ascending.
-fn distinct_sorted(scenario: &[i64]) -> Vec<i64> {
-    let mut ids = scenario.to_vec();
-    ids.sort_unstable();
-    ids.dedup();
-    ids
-}
-
-/// Every scenario of a gridfm dataset as one [`ScenarioSet`] over shared
-/// element identities: each scenario's network reuses the first scenario's
-/// table allocation wherever the rebuilt table is equal, so unchanged
-/// topology and parameters are stored once and only the tables a scenario
-/// actually changes are held per scenario. Scenario ids are the dataset's
-/// `scenario` values, ascending; the diagnostics are every scenario's read
-/// findings in that order.
-///
-/// The current gridfm profile is raw snapshot data that names no solved
-/// calculation, so the set is network data — never a solution set.
-///
-/// [`ScenarioSet`]: powerio_core::ScenarioSet
-///
-/// # Errors
-/// Propagates [`read_gridfm_scenarios`], plus a scenario identity the set
-/// rejects.
-pub fn read_gridfm_scenario_set(
-    dir: impl AsRef<Path>,
-) -> Result<(
-    powerio_core::ScenarioSet<BalancedNetwork>,
-    Vec<powerio_core::Diagnostic>,
-)> {
-    let reads = read_gridfm_scenarios(dir)?;
-    let mut diagnostics = Vec::new();
-    let mut scenarios = Vec::with_capacity(reads.len());
-    let mut donor: Option<BalancedNetwork> = None;
-    for read in reads {
-        let mut network = read.network;
-        match &donor {
-            Some(base) => network.share_equal_tables(base),
-            None => donor = Some(network.clone()),
-        }
-        diagnostics.extend(read.diagnostics);
-        let id = powerio_core::ScenarioId::new(read.scenario.to_string())?;
-        scenarios.push(powerio_core::Scenario::new(id, None, network));
-    }
-    let set = powerio_core::ScenarioSet::new(scenarios)?;
-    Ok((set, diagnostics))
-}
-
-/// The unperturbed base case: [`read_gridfm_dataset`] at `scenario = 0` (datakit's
-/// convention). There is no single "shared base" beyond a chosen scenario — bus
-/// types, branch status, and reference bus all vary per scenario — so the base
-/// case is just scenario 0.
-///
-/// # Errors
-/// Propagates [`read_gridfm_dataset`].
-pub fn gridfm_base_case(dir: impl AsRef<Path>) -> Result<GridfmRead> {
-    read_gridfm_dataset(dir, 0)
-}
-
-/// Every `bus_data` column the reader uses, concatenated across all batches once.
-/// Extracting columns up front lets a multi-scenario read reuse them rather than
-/// re-concatenating the whole table for each scenario.
-struct BusColumns {
-    scenario: Vec<i64>,
-    bus: Vec<i64>,
-    pv: Vec<i64>,
-    refc: Vec<i64>,
-    vm: Vec<f64>,
-    va: Vec<f64>,
-    vn_kv: Vec<f64>,
-    min_vm: Vec<f64>,
-    max_vm: Vec<f64>,
-    pd: Vec<f64>,
-    qd: Vec<f64>,
-    gs: Vec<f64>,
-    bs: Vec<f64>,
-}
-
-fn bus_columns(batches: &[RecordBatch]) -> Result<BusColumns> {
-    Ok(BusColumns {
-        scenario: i64_col(batches, "scenario")?,
-        bus: i64_col(batches, "bus")?,
-        pv: i64_col(batches, "PV")?,
-        refc: i64_col(batches, "REF")?,
-        vm: f64_col(batches, "Vm")?,
-        va: f64_col(batches, "Va")?,
-        vn_kv: f64_col(batches, "vn_kv")?,
-        min_vm: f64_col(batches, "min_vm_pu")?,
-        max_vm: f64_col(batches, "max_vm_pu")?,
-        pd: f64_col(batches, "Pd")?,
-        qd: f64_col(batches, "Qd")?,
-        gs: f64_col(batches, "GS")?,
-        bs: f64_col(batches, "BS")?,
-    })
-}
-
-/// Every `gen_data` column the reader uses (cost is `cp0`/`cp1`/`cp2`).
-struct GenColumns {
-    scenario: Vec<i64>,
-    bus: Vec<i64>,
-    p_mw: Vec<f64>,
-    q_mvar: Vec<f64>,
-    min_p: Vec<f64>,
-    max_p: Vec<f64>,
-    min_q: Vec<f64>,
-    max_q: Vec<f64>,
-    cp0: Vec<f64>,
-    cp1: Vec<f64>,
-    cp2: Vec<f64>,
-    in_service: Vec<i64>,
-}
-
-fn gen_columns(batches: &[RecordBatch]) -> Result<GenColumns> {
-    Ok(GenColumns {
-        scenario: i64_col(batches, "scenario")?,
-        bus: i64_col(batches, "bus")?,
-        p_mw: f64_col(batches, "p_mw")?,
-        q_mvar: f64_col(batches, "q_mvar")?,
-        min_p: f64_col(batches, "min_p_mw")?,
-        max_p: f64_col(batches, "max_p_mw")?,
-        min_q: f64_col(batches, "min_q_mvar")?,
-        max_q: f64_col(batches, "max_q_mvar")?,
-        cp0: f64_col(batches, "cp0_eur")?,
-        cp1: f64_col(batches, "cp1_eur_per_mw")?,
-        cp2: f64_col(batches, "cp2_eur_per_mw2")?,
-        in_service: i64_col(batches, "in_service")?,
-    })
-}
-
-/// Every `branch_data` column the reader uses (`Y**` / flow columns are ignored).
-struct BranchColumns {
-    scenario: Vec<i64>,
-    from_bus: Vec<i64>,
-    to_bus: Vec<i64>,
-    r: Vec<f64>,
-    x: Vec<f64>,
-    b: Vec<f64>,
-    tap: Vec<f64>,
-    shift: Vec<f64>,
-    ang_min: Vec<f64>,
-    ang_max: Vec<f64>,
-    rate_a: Vec<f64>,
-    status: Vec<i64>,
-}
-
-fn branch_columns(batches: &[RecordBatch]) -> Result<BranchColumns> {
-    Ok(BranchColumns {
-        scenario: i64_col(batches, "scenario")?,
-        from_bus: i64_col(batches, "from_bus")?,
-        to_bus: i64_col(batches, "to_bus")?,
-        r: f64_col(batches, "r")?,
-        x: f64_col(batches, "x")?,
-        b: f64_col(batches, "b")?,
-        tap: f64_col(batches, "tap")?,
-        shift: f64_col(batches, "shift")?,
-        ang_min: f64_col(batches, "ang_min")?,
-        ang_max: f64_col(batches, "ang_max")?,
-        rate_a: f64_col(batches, "rate_a")?,
-        status: i64_col(batches, "br_status")?,
-    })
-}
-
-/// The shared core: rebuild one scenario's [`BalancedNetwork`] from already-extracted
-/// columns. The columns are concatenated once by the caller and reused across
-/// scenarios, so a multi-scenario read doesn't re-copy each table per scenario.
-/// `warnings` is seeded with any manifest-level notes (e.g. a defaulted
-/// `base_mva`) and extended with the per-read fidelity notes.
-// tap == 1.0 / != 0.0 reads are exact, not approximate; the builder is one long
-// linear pass over the three tables, so the length is inherent.
-#[allow(clippy::float_cmp, clippy::too_many_lines)]
-fn build_network_from_columns(
-    bus: &BusColumns,
-    gens: &GenColumns,
-    branch: &BranchColumns,
-    scenario: i64,
-    base_mva: f64,
-    name: &str,
-    mut warnings: crate::diagnostics::Diagnostics,
-) -> Result<GridfmRead> {
-    // --- buses, loads, shunts (bus_data) ---
-    let bus_rows = scenario_rows(&bus.scenario, scenario);
-    if bus_rows.is_empty() {
-        let mut avail = bus.scenario.clone();
-        avail.sort_unstable();
-        avail.dedup();
-        return Err(powerio_tx::Error::FormatRead {
-            format: "gridfm",
-            message: format!("scenario {scenario} not present; available: {avail:?}"),
-        }
-        .into());
-    }
-
-    let bus_id = &bus.bus;
-    let pv = &bus.pv;
-    let refc = &bus.refc;
-    let vm = &bus.vm;
-    let va = &bus.va;
-    let vn_kv = &bus.vn_kv;
-    let min_vm = &bus.min_vm;
-    let max_vm = &bus.max_vm;
-    let pd = &bus.pd;
-    let qd = &bus.qd;
-    let gs = &bus.gs;
-    let bs = &bus.bs;
-
-    let mut buses = Vec::with_capacity(bus_rows.len());
-    let mut loads = Vec::new();
-    let mut shunts = Vec::new();
-    // Dense bus index -> voltage magnitude, so a generator recovers its `vg`
-    // setpoint from its bus (gridfm has no separate gen voltage column, but a
-    // generator's setpoint is its bus's regulated `Vm`).
-    let mut bus_vm: std::collections::HashMap<i64, f64> =
-        std::collections::HashMap::with_capacity(bus_rows.len());
-    for &r in &bus_rows {
-        let id = dense_bus_id(bus_id[r])?;
-        bus_vm.insert(bus_id[r], vm[r]);
-        // REF / PV / PQ one-hot; the writer guarantees exactly one set, but read
-        // defensively (REF wins, then PV, else PQ).
-        let kind = if refc[r] != 0 {
-            BusType::Ref
-        } else if pv[r] != 0 {
-            BusType::Pv
-        } else {
-            BusType::Pq
-        };
-        let mut bus = Bus::new(id, kind, vn_kv[r]);
-        bus.vm = vm[r];
-        bus.va = va[r];
-        bus.vmax = max_vm[r];
-        bus.vmin = min_vm[r];
-        bus.area = 0;
-        bus.zone = 0;
-        buses.push(bus);
-        if pd[r] != 0.0 || qd[r] != 0.0 {
-            loads.push(Load::new(id, pd[r], qd[r]));
-        }
-        // Undo the writer's `/ base_mva` (gridfm.rs:487) to recover MW/MVAr at V=1.
-        if gs[r] != 0.0 || bs[r] != 0.0 {
-            shunts.push(Shunt::new(id, gs[r] * base_mva, bs[r] * base_mva));
-        }
-    }
-
-    // --- generators (gen_data) ---
-    let gen_rows = scenario_rows(&gens.scenario, scenario);
-    require_scenario_block(&gens.scenario, scenario, &gen_rows, "gen_data")?;
-    let g_bus = &gens.bus;
-    let p_mw = &gens.p_mw;
-    let q_mvar = &gens.q_mvar;
-    let min_p = &gens.min_p;
-    let max_p = &gens.max_p;
-    let min_q = &gens.min_q;
-    let max_q = &gens.max_q;
-    let cp0 = &gens.cp0;
-    let cp1 = &gens.cp1;
-    let cp2 = &gens.cp2;
-    let g_in = &gens.in_service;
-
-    let mut generators = Vec::with_capacity(gen_rows.len());
-    for &r in &gen_rows {
-        // Any nonzero coefficient is a real polynomial cost. An all-zero triple is
-        // ambiguous in the schema — a generator with no cost, a genuine zero
-        // polynomial cost, or a piecewise/cubic+ cost the writer couldn't represent
-        // (all written as `(0, 0, 0)`) — so it reads back as `None` (see warnings).
-        let cost = if cp0[r] != 0.0 || cp1[r] != 0.0 || cp2[r] != 0.0 {
-            Some(GenCost::new(2, 0.0, 0.0, vec![cp2[r], cp1[r], cp0[r]]))
-        } else {
-            None
-        };
-        let mut generator = Generator::new(dense_bus_id(g_bus[r])?);
-        generator.pg = p_mw[r];
-        generator.qg = q_mvar[r];
-        generator.pmax = max_p[r];
-        generator.pmin = min_p[r];
-        generator.qmax = max_q[r];
-        generator.qmin = min_q[r];
-        // The schema has no gen vg; recover the setpoint from the bus's Vm
-        // (falls back to 1.0 only if the gen references an absent bus, which
-        // `validate()` below then rejects).
-        generator.vg = bus_vm.get(&g_bus[r]).copied().unwrap_or(1.0);
-        generator.mbase = base_mva;
-        generator.in_service = g_in[r] != 0;
-        generator.cost = cost;
-        generators.push(generator);
-    }
-
-    // --- branches (branch_data); Y** and pf/qf/pt/qt are ignored ---
-    let br_rows = scenario_rows(&branch.scenario, scenario);
-    require_scenario_block(&branch.scenario, scenario, &br_rows, "branch_data")?;
-    let from_bus = &branch.from_bus;
-    let to_bus = &branch.to_bus;
-    let r_col = &branch.r;
-    let x_col = &branch.x;
-    let b_col = &branch.b;
-    let tap = &branch.tap;
-    let shift = &branch.shift;
-    let ang_min = &branch.ang_min;
-    let ang_max = &branch.ang_max;
-    let rate_a = &branch.rate_a;
-    let br_status = &branch.status;
-
-    let mut branches = Vec::with_capacity(br_rows.len());
-    // The writer stores the *effective* tap (`Branch::effective_tap`), so a line
-    // (raw tap 0) lands as 1.0. Map unit tap + no shift back to the raw `tap == 0`
-    // line convention, otherwise every line reads as a transformer
-    // (`is_transformer` keys off `tap != 0`) and a read→write to a format that
-    // splits lines from transformers (PSS/E, PowerWorld) misclassifies them. A
-    // genuine unity-ratio, zero-shift transformer is electrically identical to a
-    // line and is (unavoidably) read as one.
-    let mut unit_tap_lines = 0usize;
-    for &row in &br_rows {
-        let shift_v = shift[row];
-        let tap_out = if tap[row] == 1.0 && shift_v == 0.0 {
-            unit_tap_lines += 1;
-            0.0
-        } else {
-            tap[row]
-        };
-        let mut branch = Branch::new(
-            dense_bus_id(from_bus[row])?,
-            dense_bus_id(to_bus[row])?,
-            r_col[row],
-            x_col[row],
-        );
-        branch.b = b_col[row];
-        branch.rate_a = rate_a[row];
-        branch.tap = tap_out;
-        branch.shift = shift_v;
-        branch.in_service = br_status[row] != 0;
-        branch.angmin = ang_min[row];
-        branch.angmax = ang_max[row];
-        branches.push(branch);
-    }
-
-    let mut net = BalancedNetwork::new(name, base_mva);
-    *net.buses_mut() = buses;
-    *net.loads_mut() = loads;
-    *net.shunts_mut() = shunts;
-    *net.branches_mut() = branches;
-    *net.generators_mut() = generators;
-    *net.source_format_mut() = SourceFormat::Gridfm;
-    net.validate()?;
-
-    // --- fidelity warnings: what the gridfm schema couldn't carry back ---
-    warnings.push(
-        &codes::READ_GRIDFM_VALUE_INFERRED,
-        format!(
-            "synthesized bus ids 1..={}; original bus ids are not stored in a gridfm dataset, \
-         so a written case is renumbered",
-            net.buses().len()
-        ),
-    );
-    if !net.loads().is_empty() {
-        warnings.push(
-            &codes::READ_GRIDFM_VALUE_COLLAPSED,
-            format!(
-                "folded nodal load into {} synthetic per-bus Load(s); per-load granularity is \
-             not recoverable",
-                net.loads().len()
-            ),
-        );
-    }
-    if !net.shunts().is_empty() {
-        warnings.push(
-            &codes::READ_GRIDFM_VALUE_COLLAPSED,
-            format!(
-                "folded nodal shunts into {} synthetic per-bus Shunt(s); per-shunt granularity \
-             is not recoverable",
-                net.shunts().len()
-            ),
-        );
-    }
-    if unit_tap_lines > 0 {
-        warnings.push(&codes::READ_GRIDFM_ELEMENT_RELABELED, format!(
-            "{unit_tap_lines} branch(es) had unit effective tap and no phase shift and were read \
-             as lines (raw tap 0); a unity-ratio, zero-shift transformer in the source is \
-             indistinguishable from a line and is read as one (the power flow is identical)"
-        ));
-    }
-    let no_cost_gens = net.generators().iter().filter(|g| g.cost.is_none()).count();
-    if no_cost_gens > 0 {
-        warnings.push(&codes::READ_GRIDFM_FIELD_DROPPED, format!(
-            "{no_cost_gens} generator(s) read with no cost: an all-zero cost triple in the dataset \
-             is the writer's encoding for a generator with no cost, a genuine zero polynomial \
-             cost, or a piecewise/cubic+ cost it couldn't represent — indistinguishable on read"
-        ));
-    }
-    warnings.push(
-        &codes::READ_GRIDFM_FIELD_DROPPED,
-        "HVDC, storage, areas/zones, bus names, rate_b/rate_c, generator mbase/ramp limits, \
-         and startup/shutdown costs are absent from the gridfm schema",
-    );
-
-    Ok(GridfmRead {
-        network: net,
-        scenario,
-        warnings: warnings.lines(),
-        diagnostics: warnings.into_records(),
-    })
-}
-
-// --- reader helpers --------------------------------------------------------
-
-/// Resolve the directory holding the gridfm parquet files, accepting (in order)
-/// the leaf `raw/` dir itself, a `<case>/` dir with a `raw/` child, or a parent
-/// dir with exactly one `*/raw/` child.
-fn resolve_raw_dir(dir: &Path) -> Result<PathBuf> {
-    let has_bus = |d: &Path| d.join("bus_data.parquet").is_file();
-    if has_bus(dir) {
-        return Ok(dir.to_path_buf());
-    }
-    let nested = dir.join("raw");
-    if has_bus(&nested) {
-        return Ok(nested);
-    }
-    // Parent-of-<case> fallback: scan for exactly one `*/raw/bus_data.parquet`.
-    // Surface read_dir / entry IO errors (missing dir, permissions) rather than
-    // masking them as "no dataset found".
-    let mut matches: Vec<PathBuf> = Vec::new();
-    let entries = std::fs::read_dir(dir).map_err(|e| powerio_tx::Error::FormatRead {
-        format: "gridfm",
-        message: format!("reading directory {}: {e}", dir.display()),
-    })?;
-    for entry in entries {
-        let entry = entry.map_err(|e| powerio_tx::Error::FormatRead {
-            format: "gridfm",
-            message: format!("reading an entry of {}: {e}", dir.display()),
-        })?;
-        let raw = entry.path().join("raw");
-        if has_bus(&raw) {
-            matches.push(raw);
-        }
-    }
-    match matches.len() {
-        1 => Ok(matches.pop().expect("len checked")),
-        0 => Err(powerio_tx::Error::FormatRead {
-            format: "gridfm",
-            message: format!(
-                "no gridfm dataset under {}; expected bus_data.parquet in the directory, a \
-                 raw/ child, or a single <case>/raw/ child",
-                dir.display()
-            ),
-        }
-        .into()),
-        n => Err(powerio_tx::Error::FormatRead {
-            format: "gridfm",
-            message: format!(
-                "{n} gridfm datasets under {}; point at the specific <case>/raw directory",
-                dir.display()
-            ),
-        }
-        .into()),
-    }
-}
-
-/// One `READ.GRIDFM.VALUE_DEFAULTED` note, for a manifest the reader could not
-/// use.
-fn defaulted_meta(message: impl Into<String>) -> crate::diagnostics::Diagnostics {
-    let mut diagnostics = crate::diagnostics::Diagnostics::new();
-    diagnostics.push(&codes::READ_GRIDFM_VALUE_DEFAULTED, message);
-    diagnostics
-}
-
-/// Read `base_mva` and the case name from `raw/gridfm_meta.json`. A missing or
-/// unreadable manifest is not fatal — a bare prediction may lack one — so default
-/// `base_mva` to 100, derive the name from the `<case>/raw` parent, and warn.
-fn read_meta(raw: &Path) -> (f64, String, crate::diagnostics::Diagnostics) {
-    let case_from_path = || {
-        raw.parent()
-            .and_then(Path::file_name)
-            .and_then(|s| s.to_str())
-            .map_or_else(|| "gridfm".to_string(), str::to_string)
-    };
-    let text = match std::fs::read_to_string(raw.join("gridfm_meta.json")) {
-        Ok(text) => text,
-        // Not just "not found": surface the underlying error (permissions, non-UTF-8,
-        // etc.) so a dataset problem is diagnosable, while still defaulting base_mva.
-        Err(e) => {
-            return (
-                100.0,
-                case_from_path(),
-                defaulted_meta(format!(
-                    "gridfm_meta.json could not be read ({e}); base_mva defaulted to 100"
-                )),
-            );
-        }
-    };
-    let Ok(meta) = serde_json::from_str::<serde_json::Value>(&text) else {
-        return (
-            100.0,
-            case_from_path(),
-            defaulted_meta("gridfm_meta.json is not valid JSON; base_mva defaulted to 100"),
-        );
-    };
-    let name = meta
-        .get("case_name")
-        .and_then(serde_json::Value::as_str)
-        .map_or_else(case_from_path, str::to_string);
-    let mut warnings = crate::diagnostics::Diagnostics::new();
-    // base_mva must be a positive, finite number; anything else (absent, zero,
-    // negative, NaN) defaults to 100 with a note — shunt recovery scales by it.
-    let base = match meta.get("base_mva").and_then(serde_json::Value::as_f64) {
-        Some(b) if b.is_finite() && b > 0.0 => b,
-        _ => {
-            warnings.push(
-                &codes::READ_GRIDFM_VALUE_DEFAULTED,
-                "gridfm_meta.json has no usable base_mva (absent or not a positive number); \
-                 defaulted to 100",
-            );
-            100.0
-        }
-    };
-    (base, name, warnings)
-}
-
-/// Open a parquet file and collect every record batch (one for powerio-written
-/// data, possibly several for an externally-written prediction).
-fn read_parquet(path: &Path) -> Result<Vec<RecordBatch>> {
-    let file = std::fs::File::open(path).map_err(|e| powerio_tx::Error::FormatRead {
-        format: "gridfm",
-        message: format!("opening {}: {e}", path.display()),
-    })?;
-    let reader = ParquetRecordBatchReaderBuilder::try_new(file)
-        .and_then(ParquetRecordBatchReaderBuilder::build)
-        .map_err(|e| powerio_tx::Error::FormatRead {
-            format: "gridfm",
-            message: format!("reading {}: {e}", path.display()),
-        })?;
-    reader
-        .collect::<std::result::Result<Vec<_>, _>>()
-        .map_err(|e| {
-            Error::Core(powerio_tx::Error::FormatRead {
-                format: "gridfm",
-                message: format!("decoding {}: {e}", path.display()),
-            })
-        })
-}
-
-/// Row indices whose `scenario` column equals `scenario`, in table order.
-fn scenario_rows(scen: &[i64], scenario: i64) -> Vec<usize> {
-    scen.iter()
-        .enumerate()
-        .filter_map(|(i, &s)| (s == scenario).then_some(i))
-        .collect()
-}
-
-/// Guard a gen/branch table: empty `rows` is fine when the whole table is empty
-/// (a case legitimately has no generators, or no branches), but if the table
-/// holds rows for *other* scenarios yet none for this one, the dataset is partial
-/// or corrupt and would silently yield a wrong-but-valid network — error instead.
-fn require_scenario_block(
-    scen_col: &[i64],
-    scenario: i64,
-    rows: &[usize],
-    table: &str,
-) -> Result<()> {
-    if rows.is_empty() && !scen_col.is_empty() {
-        return Err(powerio_tx::Error::FormatRead {
-            format: "gridfm",
-            message: format!(
-                "scenario {scenario} has no {table} rows, but the table holds {} row(s) for other \
-                 scenarios — a partial or corrupt dataset",
-                scen_col.len()
-            ),
-        }
-        .into());
-    }
-    Ok(())
-}
-
-/// A dense `[0, n)` parquet index → 1-based [`BusId`]. Errors on a negative index.
-fn dense_bus_id(v: i64) -> Result<BusId> {
-    let idx = usize::try_from(v).map_err(|_| powerio_tx::Error::FormatRead {
-        format: "gridfm",
-        message: format!("negative dense bus index {v}"),
-    })?;
-    Ok(BusId(idx + 1))
-}
-
-/// Look up a named column, erroring if absent.
-fn column<'a>(b: &'a RecordBatch, name: &str) -> Result<&'a ArrayRef> {
-    b.column_by_name(name).ok_or_else(|| {
-        Error::Core(powerio_tx::Error::FormatRead {
-            format: "gridfm",
-            message: format!("missing column `{name}`"),
-        })
-    })
-}
-
-/// Concatenate a named non-null `Int64` column across all batches.
-fn i64_col(batches: &[RecordBatch], name: &str) -> Result<Vec<i64>> {
-    let mut out = Vec::with_capacity(batches.iter().map(RecordBatch::num_rows).sum());
-    for b in batches {
-        let arr = column(b, name)?;
-        let col = arr.as_any().downcast_ref::<Int64Array>().ok_or_else(|| {
-            powerio_tx::Error::FormatRead {
-                format: "gridfm",
-                message: format!("column `{name}` is not Int64"),
-            }
-        })?;
-        if col.null_count() > 0 {
-            return Err(powerio_tx::Error::FormatRead {
-                format: "gridfm",
-                message: format!("column `{name}` has nulls"),
-            }
-            .into());
-        }
-        out.extend_from_slice(col.values());
-    }
-    Ok(out)
-}
-
-/// Concatenate a named non-null `Float64` column across all batches.
-fn f64_col(batches: &[RecordBatch], name: &str) -> Result<Vec<f64>> {
-    let mut out = Vec::with_capacity(batches.iter().map(RecordBatch::num_rows).sum());
-    for b in batches {
-        let arr = column(b, name)?;
-        let col = arr.as_any().downcast_ref::<Float64Array>().ok_or_else(|| {
-            powerio_tx::Error::FormatRead {
-                format: "gridfm",
-                message: format!("column `{name}` is not Float64"),
-            }
-        })?;
-        if col.null_count() > 0 {
-            return Err(powerio_tx::Error::FormatRead {
-                format: "gridfm",
-                message: format!("column `{name}` has nulls"),
-            }
-            .into());
-        }
-        out.extend_from_slice(col.values());
-    }
-    Ok(out)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::network::SourceFormat;
     use crate::network::{Branch, BranchCharging, Bus, BusId, BusType, Generator};
     use arrow::array::{Float64Array, Int64Array};
     use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
+    use powerio::gridfm::{
+        gridfm_base_case, read_gridfm_dataset, read_gridfm_network, read_gridfm_scenario_set,
+        read_gridfm_scenarios,
+    };
 
     const BUS_COLS: &[&str] = &[
         "scenario",
@@ -2967,7 +2232,15 @@ mod tests {
         // fingerprint with no disk I/O.
         let net = case14();
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let read = read_gridfm_network(&tables, 0, net.base_mva(), net.name()).unwrap();
+        let read = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            0,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap();
         assert_fingerprint_close(&read.network, &net);
     }
 
@@ -2980,7 +2253,15 @@ mod tests {
         assert!(want_b.abs() > 1.0, "fixture should have a real shunt");
 
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let read = read_gridfm_network(&tables, 0, net.base_mva(), net.name()).unwrap();
+        let read = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            0,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap();
         let got_b: f64 = read.network.shunts().iter().map(|s| s.b).sum();
         assert!(
             (got_b - want_b).abs() < 1e-9,
@@ -3044,14 +2325,22 @@ mod tests {
     fn read_missing_scenario_errors() {
         let net = case14();
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let err = read_gridfm_network(&tables, 99, net.base_mva(), net.name()).unwrap_err();
+        let err = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            99,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap_err();
         assert!(
             matches!(
                 err,
-                Error::Core(powerio_tx::Error::FormatRead {
+                powerio_tx::Error::FormatRead {
                     format: "gridfm",
                     ..
-                })
+                }
             ),
             "got {err:?}"
         );
@@ -3065,10 +2354,10 @@ mod tests {
         assert!(
             matches!(
                 err,
-                Error::Core(powerio_tx::Error::FormatRead {
+                powerio_tx::Error::FormatRead {
                     format: "gridfm",
                     ..
-                })
+                }
             ),
             "got {err:?}"
         );
@@ -3078,10 +2367,10 @@ mod tests {
         assert!(
             matches!(
                 err,
-                Error::Core(powerio_tx::Error::FormatRead {
+                powerio_tx::Error::FormatRead {
                     format: "gridfm",
                     ..
-                })
+                }
             ),
             "got {err:?}"
         );
@@ -3118,7 +2407,15 @@ mod tests {
     fn read_surfaces_fidelity_warnings() {
         let net = case14();
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let read = read_gridfm_network(&tables, 0, net.base_mva(), net.name()).unwrap();
+        let read = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            0,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap();
         assert!(!read.warnings.is_empty());
         assert!(
             read.warnings
@@ -3139,7 +2436,15 @@ mod tests {
         // (not the old hard-coded 1.0).
         let net = case14();
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let read = read_gridfm_network(&tables, 0, net.base_mva(), net.name()).unwrap();
+        let read = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            0,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap();
         for g in read.network.generators() {
             let bus = read
                 .network
@@ -3183,7 +2488,15 @@ mod tests {
         );
 
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let read = read_gridfm_network(&tables, 0, net.base_mva(), net.name()).unwrap();
+        let read = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            0,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap();
         let read_lines = read
             .network
             .branches()
@@ -3223,7 +2536,15 @@ mod tests {
             vec![branch(1, 2, 0.01, 0.1)],
         );
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let read = read_gridfm_network(&tables, 0, net.base_mva(), net.name()).unwrap();
+        let read = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            0,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap();
         assert!(read.network.generators().is_empty());
         assert_eq!(read.network.branches().len(), 1);
     }
@@ -3242,7 +2563,15 @@ mod tests {
         net.generators_mut()
             .push(gen_at(1, gencost(2, 3, vec![0.0, 0.0, 0.0])));
         let tables = gridfm_record_batches_single(&net, 0, &GridfmOptions::default()).unwrap();
-        let read = read_gridfm_network(&tables, 0, net.base_mva(), net.name()).unwrap();
+        let read = read_gridfm_network(
+            &tables.bus,
+            &tables.generator,
+            &tables.branch,
+            0,
+            net.base_mva(),
+            net.name(),
+        )
+        .unwrap();
         assert!(
             read.network.generators()[0].cost.is_none(),
             "all-zero cost should read back as None"
@@ -3253,25 +2582,6 @@ mod tests {
                 .any(|w| w.contains("read with no cost")),
             "expected the no-cost ambiguity warning, got {:?}",
             read.warnings
-        );
-    }
-
-    #[test]
-    fn require_scenario_block_flags_partial_tables() {
-        // Empty table → ok (a legitimately element-less case). Present → ok. Absent
-        // from a non-empty table → error (the partial/corrupt dataset Copilot flagged).
-        assert!(require_scenario_block(&[], 0, &[], "gen_data").is_ok());
-        assert!(require_scenario_block(&[0, 0, 1], 0, &[0, 1], "gen_data").is_ok());
-        let err = require_scenario_block(&[0, 0], 1, &[], "branch_data").unwrap_err();
-        assert!(
-            matches!(
-                err,
-                Error::Core(powerio_tx::Error::FormatRead {
-                    format: "gridfm",
-                    ..
-                })
-            ),
-            "got {err:?}"
         );
     }
