+++ powerio-matrix/src/lib.rs
@@ -46,7 +46,6 @@ pub use powerio_tx::{
     write_powermodels_json, write_powerworld, write_psse, write_pypsa_csv_folder,
 };
 
-mod collect;
 pub mod diagnostics;
 pub mod error;
 pub use error::{ElementCounts, Error, Result, ScenarioMismatch};
@@ -92,10 +91,6 @@ pub use pipeline::{
 
 #[cfg(feature = "gridfm")]
 pub use io::gridfm::{
-    GridfmOptions, GridfmOutputs, GridfmRead, GridfmSnapshot, GridfmTables, gridfm_base_case,
-    gridfm_record_batches, gridfm_record_batches_single, gridfm_scenario_ids, numbered_snapshots,
-    read_gridfm_dataset, read_gridfm_network, read_gridfm_scenario_set, read_gridfm_scenarios,
-    write_gridfm_batch, write_gridfm_dataset,
+    GridfmOptions, GridfmOutputs, GridfmSnapshot, GridfmTables, gridfm_record_batches,
+    gridfm_record_batches_single, numbered_snapshots, write_gridfm_batch, write_gridfm_dataset,
 };
-#[cfg(feature = "gridfm")]
-pub use io::{dataset_scenario_ids, read_dataset_dir};
