+++ powerio/Cargo.toml
@@ -27,11 +27,15 @@ powerio-matrix = { workspace = true, optional = true }
 powerio-prob = { workspace = true }
 powerio-tx = { workspace = true }
 schemars = { workspace = true, optional = true }
+bytes = "1.9"
 serde = { workspace = true }
 serde_json = { workspace = true }
+arrow = { workspace = true }
+parquet = { version = "59", default-features = false, features = ["arrow", "snap"] }
 thiserror = "2"
 
 [dev-dependencies]
+powerio-matrix = { workspace = true, features = ["gridfm"] }
 tempfile = "3"
 
 [[example]]
